+++ packages/client/src/assets/img/errorZombies/index.ts
