--- packages/client/src/ui/flat_button/style.module.scss
@@ -1,48 +1,0 @@
-.button {
-  position: relative;
-  width: 280px;
-  height: 37px;
-  border-radius: 8px;
-  color: var(--white-color);
-  background-color: var(--primary-color);
-  transition:
-    box-shadow var(--transition-bezier),
-    background-color var(--transition-bezier);
-
-  &--transparent {
-    background-color: transparent;
-    color: var(--primary-color);
-    transition: text-shadow var(--transition-bezier);
-  }
-
-  &--dangerous {
-    background-color: var(--danger-color);
-  }
-
-  &:hover {
-    box-shadow: 0px 0px 5px 0px var(--shadow-color-strong);
-
-    &.button--transparent {
-      box-shadow: 0px 0px 0px 0px var(--black-color);
-      text-shadow: 0px 0px 0.7px var(--shadow-color-strong);
-    }
-  }
-
-  &:active {
-    box-shadow: 0px 0px 1px 0px var(--shadow-color-strong);
-
-    &.button--transparent {
-      box-shadow: 0px 0px 0px 0px var(--black-color);
-      text-shadow: 0px 0px 0.2px var(--shadow-color-strong);
-    }
-  }
-
-  &:disabled {
-    background-color: var(--border-color);
-    cursor: not-allowed;
-
-    &:hover {
-      box-shadow: 0px 0px 0px 0px var(--black-color);
-    }
-  }
-}
