+++ packages/client/src/ui/title/style.module.scss
@@ -1,5 +1,5 @@
 .title {
   font-size: 24px;
   line-height: 28px;
-  color: var(--text-color);
+  color: var(--text);
 }
