+++ packages/client/src/ui/zombie_error/style.module.scss
@@ -41,7 +41,7 @@
   }
 
   &__text {
-    color: var(--text-color);
+    color: var(--text);
   }
 }
